feat: 예외 처리를 하기 위해 RequestHandler의 run() 메서드에서 HttpRequest, HttpResponse 객체 생성 후 수행하는 로직을 FrontController 클래스 생성 후 FrontController의 process() 메서드로 이동
- FrontController 클래스 생성
- WebServer 클래스의 main() 메서드에서 FrontController 객체 생성 후 RequestHandler 생성자의 인자로 frontController를 넘겨줌
- RequestHandler의 private final 멤버변수로 FrontController 추가
- RequestHandler의 run() 메서드에서 frontController.process() 메서드 호출
- frontController.process() 메서드
    - 매핑된 컨트롤러를 찾지 못했을 때 발생하는 ControllerNotFoundException, 잘못된 URL 요청을 보냈을 때 발생하는 UrlNotFoundException, 파일을 찾지 못했을 때 FileNotFoundException을 처리하는 try-catch 문 추가
