+++ src/main/java/webserver/WebServer.java
@@ -22,10 +22,11 @@ public class WebServer {
         try (ServerSocket listenSocket = new ServerSocket(port)) {
             logger.info("Web Application Server started {} port.", port);
 
-            // 클라이언트가 연결될때까지 대기한다.
+            FrontController frontController = new FrontController();
             Socket connection;
+            // 클라이언트가 연결될때까지 대기한다.
             while ((connection = listenSocket.accept()) != null) {
-                Thread thread = new Thread(new RequestHandler(connection));
+                Thread thread = new Thread(new RequestHandler(connection, frontController));
                 thread.start();
             }
         }
